Divide SSO spec into separate files (for each of the providers)
